Fix tests + improve testing script
Now, the tests are supposed to be run from the root directory of the
repo with ./tests/run_all_tests_debug.sh. It will automatically re-build
the compiler, and will only show the output/logs of the test if they
fail.

The logs are also not dumped to stdout, and are instead put into a file
separately which is shown in the output.

All test output/default built paths are all now in the `./build/` folder
since frederik seems to show a lack of regard for good git hygiene
(angry face), and this way they will get ignored by default.